Refatorando os códigos referentes ao Investimento
